+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: mharissa <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/04 19:12:49 by mharissa          #+#    #+#             */
-/*   Updated: 2019/09/11 18:16:42 by mharissa         ###   ########.fr       */
+/*   Updated: 2019/09/24 20:40:00 by mharissa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	size_t	i;
